+++ Src/FileTextStats.h
@@ -13,10 +13,10 @@
  */
 struct FileTextStats
 {
-	unsigned ncrs; /**< Count of MAC (CR-byte) EOLs. */
-	unsigned nlfs; /**< Count of Unix (LF-byte) EOLs. */
-	unsigned ncrlfs; /**< Count of DOS (CR+LF-bytes) EOLs. */
-	unsigned nzeros; /**< Count of zero-bytes. */
+	__int64 ncrs; /**< Count of MAC (CR-byte) EOLs. */
+	__int64 nlfs; /**< Count of Unix (LF-byte) EOLs. */
+	__int64 ncrlfs; /**< Count of DOS (CR+LF-bytes) EOLs. */
+	__int64 nzeros; /**< Count of zero-bytes. */
 	FileTextStats() { clear(); }
 	void clear() { ncrs = nlfs = ncrlfs = nzeros = 0; }
 };
